Added slider
