--- src/dropdown.css
@@ -1,13 +1,0 @@
-.down{
-    display: flex;
-    justify-content: space-between;
-    justify-items: center;
-    padding-top: 40px;
-}
-
-.drop-btn{
-    padding: 0;
-    border: none;
-    background: none;
-}
-
