+++ src/screens/home/apartaments.css
@@ -1,0 +1,32 @@
+.section-one-container {
+    width: 100%;
+    padding: 60px 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 0;
+    flex-direction: column;
+}
+
+.text-text {
+    width: 1018px;
+    height: 111px;
+    font-style: normal;
+    font-weight: 700;
+    font-size: 20px;
+    line-height: 26px;
+    color: #000000;
+}
+
+.title-first-section {
+    font-style: normal;
+    font-weight: 700;
+    font-size: 40px;
+    line-height: 52px;
+    text-align: center;
+    letter-spacing: 0.05em;
+    color: #FFB701;
+    margin: 30px;
+}
+
+
