+++ src/screens/home/components/slides.css
@@ -2,8 +2,6 @@
     display: flex;
     justify-content: space-between;
     justify-items: center;
-    padding-top: 40px;
-
 }
 
 .container {
@@ -12,26 +10,24 @@
     display: flex;
     justify-content: space-between;
     padding: 20px;
-
     background: #FFB701;
     box-shadow: 5px 5px 70px rgba(0, 0, 0, 0.25), inset 0px 1px 10px rgba(0, 0, 0, 0.25);
     border-radius: 10px;
-
 }
 
 .prev, .next {
     align-items: center;
     display: flex;
+    padding-right: 70px;
+    padding-left: 70px;
 }
-
 
 .txt {
     font-style: normal;
     font-weight: 700;
     font-size: 16px;
     line-height: 21px;
-    padding-left: 18px;
-    padding-top: 30px;
+    padding-top: 46px;
     letter-spacing: 0.05em;
     color: #000000;
 }
@@ -42,16 +38,16 @@
     font-size: 16px;
     line-height: 21px;
     letter-spacing: 0.05em;
-    display: flex;
     color: #000000;
-    padding-top: 38px;
-    padding-left: 18px;
+    padding-top: 36px;
+    display: flex;
 }
 
-.countdown {
+.countdown-first {
     font-weight: bold;
     font-size: 30px;
     color: #FFFFFF;
+    margin-left: 16px;
 }
 
 .left-part {
@@ -80,7 +76,6 @@
 }
 
 .one {
-    /*font-weight: 700;*/
     line-height: 21px;
     letter-spacing: 0.05em;
     text-decoration-line: underline;
@@ -107,10 +102,10 @@
 .button-container {
     width: 100%;
     height: 100px;
-    border: 1px solid #337AB7;
     display: flex;
     flex-direction: column;
     justify-content: space-around;
+    padding-top: 50px;
 
 }
