+++ src/screens/home/homepage.css
@@ -1,7 +1,6 @@
 .homepage-wrapper {
     background: #FFB701;
     width: 100%;
-    height: 88vh;
     display: flex;
     align-items: center;
     justify-content: center;
